Fix scripts/build modules=...
Fix building the "httpserver" module, needed to use build modules=..,
instead of image=...

There's one catch though - the httpserver assumes we previously ran
in the top-level OSv directory

    (cd java; mvn package -q -DskipTests=true)

(This was run in the old build.mk)

This was a really bad assumption, because what this produces
(java/jolokia-agent/target/) has *nothing* to do with the main makefile.
If we really need this, we should have a separate module that does this
mvn thing, and the httpserver module should require it.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>