Cleanup modules/tests/Makefile
The Java compilation is an unused relic, so we can remove it (there's
a separate java-tests instead). Add "-O2" to the tests' compilation,
otherwise some of the tests will be incredibly slow.

scripts/test.py, running all the tests, succeeds. There are still some
ZFS tests I didn't wire in, but it's already in a good shape.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>