build: Improve image size setting
Fix the command
	scripts/build image=rogue fs_size_mb=1234

To generate a virtual disk exactly 1234MB (1234*1024*1024 bytes)
in size. Previously its size was mistakenly the sum of that size
and the kernel size, though only the part up to 1234MB was
actually used by ZFS.

Also support a new specification of exact number of bytes:

	scripts/build image=rogue fs_size=100000000

This number is rounded down to the nearest multiple of 512.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>